requirements/base.txt
# Core Django
Django==5.1.4
djangorestframework==3.16.1

# Async & WebSockets
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1

# Database
psycopg2-binary==2.9.11

# Security & Auth
python-dotenv==1.2.1
django-cors-headers==4.9.0

# Media
Pillow==12.1.0

# Redis
redis==7.1.0
